+++ src/test/java/de/unistuttgart/overworldbackend/PlayerStatisticControllerTest.java
@@ -38,7 +38,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @Transactional
 @SpringBootTest
 @Testcontainers
-class PlayerInitialDataStatisticControllerTest {
+class PlayerStatisticControllerTest {
 
     @Container
     public static PostgreSQLContainer postgresDB = new PostgreSQLContainer("postgres:14-alpine")
